LS: Move `get_diagnostics` out of the main module (#5316)
