+++ crates/cairo-lang-language-server/src/lang/diagnostics/mod.rs
@@ -1,0 +1,1 @@
+pub mod lsp;
