+++ crates/cairo-lang-language-server/src/lang/mod.rs
@@ -1,1 +1,2 @@
+pub mod diagnostics;
 pub mod lsp;
